Refactor to remove mutability from send and recv implementtion.
* Fixes a bug with polling which might spuriously return negative result
when there is contention on the channel.
* Notifying another domain from within a critical section is unsafe in
general. Notify blocks until the target domain is out of the critical
section. If two domains are notifying each other from within critical
section, then the program deadlocks. However, in the channel
implementation, if a domain is currently not waiting, then no other
domain will notify it. Hence, it is correct to notify from within the
critical section.