Fix Task.parallel_for_reduce on empty loop
`Task.parallel_for_reduce` gives the wrong result for a the corner-case
of an empty loop with `start:0` and `finish:-1`.

For a simple (in this case: empty) loop that adds an array of 1s,
it returns `1` rather than the expected `init = 0`.

The current code builds on a reasonable assumption of non-empty parallel loops.
The PR tests for the corner-case only once per invocation ("at top-level")
and hence should not affect performance.